+++ src/tunnel/tun_interface.hpp
@@ -30,19 +30,21 @@ class tun_interface
 {
 public:
 
+    static std::unique_ptr<tun_interface> make_tun_interface(
+        boost::asio::io_service& io,
+        const std::string& wanted_device_name,
+        std::error_code& error);
+
+public:
+
     using io_handler = std::function<void(const std::error_code&,uint32_t)>;
 
     // constructor
     tun_interface(boost::asio::io_service& io,
-                  int tun_fd,
-                  const std::string& devname);
+                  int file_descriptor,
+                  const std::string& device_name);
 
     ~tun_interface();
-
-    static std::unique_ptr<tun_interface> make_tun_interface(
-        boost::asio::io_service& io,
-        std::string& devname,
-        std::error_code& error);
 
     void set_ipv4(const std::string& address, std::error_code& error);
     uint32_t mtu(std::error_code& error) const;
@@ -54,20 +56,19 @@ public:
     bool is_down(std::error_code& error) const;
     void down(std::error_code& error);
 
-    std::string name() const;
+    std::string device_name() const;
 
     // Read buffers / packets sent to this interface from the kernel.
     // That is, "outbound traffic".
-    void async_read(std::vector<uint8_t>& buffer, io_handler callback);
+    void async_read(uint8_t* data, uint32_t size, io_handler callback);
 
     // Write buffers / packets to the kernel from this interface.
     // That is, "inbound traffic".
-    void async_write(const std::vector<uint8_t>& buffer,
-                     io_handler callback);
+    void async_write(const uint8_t* data, uint32_t size, io_handler callback);
 
     // Write buffers / packets to the kernel from this interface.
     // That is, "inbound traffic".
-    void write(const std::vector<uint8_t>& buffer, std::error_code& error);
+    void write(const uint8_t* data, uint32_t size, std::error_code& error);
 
     // Set default route to this interface
     void set_default_route(std::error_code& error);
@@ -78,15 +79,15 @@ public:
 private:
 
     // The dev name
-    std::string m_name;
+    std::string m_device_name;
 
     // Internal kernel socket used for ioctl calls
-    int m_kernel_socket;
+    int m_socket;
 
-    // The tun dev file descriptor
-    int m_tun_fd = -1;
+    // The tun device file descriptor
+    int m_file_descriptor = -1;
 
-    // The tun fd stream descriptor
-    boost::asio::posix::stream_descriptor m_tun_stream;
+    // The tun file descriptor stream descriptor
+    boost::asio::posix::stream_descriptor m_stream_descriptor;
 };
 }